+++ internals/constants/src/index.ts
@@ -3,3 +3,7 @@ export const PROJECT_ENV_TARGET = [
   'preview',
   'development',
 ] as const;
+
+export const LOGO = '▲' as const;
+export const NAME = 'vercel' as const;
+export const TITLE = 'Vercel' as const;