+++ packages/cli/src/util/link/repo.ts
@@ -2,15 +2,11 @@ import chalk from 'chalk';
 import pluralize from 'pluralize';
 import { homedir } from 'os';
 import { join, normalize } from 'path';
-import { normalizePath } from '@vercel/build-utils';
-import { lstat, readJSON, outputJSON, writeFile, readFile } from 'fs-extra';
+import { normalizePath, traverseUpDirectories } from '@vercel/build-utils';
+import { lstat, readJSON, outputJSON } from 'fs-extra';
 import confirm from '../input/confirm';
 import toHumanPath from '../humanize-path';
-import {
-  VERCEL_DIR,
-  VERCEL_DIR_README,
-  VERCEL_DIR_REPO,
-} from '../projects/link';
+import { VERCEL_DIR, VERCEL_DIR_REPO, writeReadme } from '../projects/link';
 import { getRemoteUrls } from '../create-git-meta';
 import link from '../output/link';
 import { emoji, prependEmoji } from '../emoji';
@@ -44,9 +40,12 @@ export interface RepoLink {
  * and returns the parsed `.vercel/repo.json` file if the repository
  * has already been linked.
  */
-export async function getRepoLink(cwd: string): Promise<RepoLink | undefined> {
+export async function getRepoLink(
+  client: Client,
+  cwd: string
+): Promise<RepoLink | undefined> {
   // Determine where the root of the repo is
-  const rootPath = await findRepoRoot(cwd);
+  const rootPath = await findRepoRoot(client, cwd);
   if (!rootPath) return undefined;
 
   // Read the `repo.json`, if this repo has already been linked
@@ -67,7 +66,7 @@ export async function ensureRepoLink(
 ): Promise<RepoLink | undefined> {
   const { output } = client;
 
-  const repoLink = await getRepoLink(cwd);
+  const repoLink = await getRepoLink(client, cwd);
   if (repoLink) {
     output.debug(`Found Git repository root directory: ${repoLink.rootPath}`);
   } else {
@@ -193,10 +192,7 @@ export async function ensureRepoLink(
     };
     await outputJSON(repoConfigPath, repoConfig, { spaces: 2 });
 
-    await writeFile(
-      join(rootPath, VERCEL_DIR, VERCEL_DIR_README),
-      await readFile(join(__dirname, 'VERCEL_DIR_README.txt'), 'utf8')
-    );
+    await writeReadme(rootPath);
 
     // update .gitignore
     const isGitIgnoreUpdated = await addToGitIgnore(rootPath);
@@ -225,45 +221,47 @@ export async function ensureRepoLink(
  * the nearest `.git/config` file is found. Returns the directory where
  * the Git config was found, or `undefined` when no Git repo was found.
  */
-export async function findRepoRoot(start: string): Promise<string | undefined> {
-  for (const current of traverseUpDirectories(start)) {
+export async function findRepoRoot(
+  client: Client,
+  start: string
+): Promise<string | undefined> {
+  const { debug } = client.output;
+  const REPO_JSON_PATH = join(VERCEL_DIR, VERCEL_DIR_REPO);
+  const GIT_CONFIG_PATH = normalize('.git/config');
+
+  for (const current of traverseUpDirectories({ start })) {
     if (current === home) {
       // Sometimes the $HOME directory is set up as a Git repo
       // (for dotfiles, etc.). In this case it's safe to say that
       // this isn't the repo we're looking for. Bail.
+      debug('Arrived at home directory');
       break;
     }
 
     // if `.vercel/repo.json` exists (already linked),
     // then consider this the repo root
-    const repoConfigPath = join(current, VERCEL_DIR, VERCEL_DIR_REPO);
+    const repoConfigPath = join(current, REPO_JSON_PATH);
     let stat = await lstat(repoConfigPath).catch(err => {
       if (err.code !== 'ENOENT') throw err;
     });
     if (stat) {
+      debug(`Found "${REPO_JSON_PATH}" - detected "${current}" as repo root`);
       return current;
     }
 
     // if `.git/config` exists (unlinked),
     // then consider this the repo root
-    const gitConfigPath = join(current, '.git/config');
+    const gitConfigPath = join(current, GIT_CONFIG_PATH);
     stat = await lstat(gitConfigPath).catch(err => {
       if (err.code !== 'ENOENT') throw err;
     });
     if (stat) {
+      debug(`Found "${GIT_CONFIG_PATH}" - detected "${current}" as repo root`);
       return current;
     }
   }
-}
 
-export function* traverseUpDirectories(start: string) {
-  let current: string | undefined = normalize(start);
-  while (current) {
-    yield current;
-    // Go up one directory
-    const next = join(current, '..');
-    current = next === current ? undefined : next;
-  }
+  debug('Aborting search for repo root');
 }
 
 function sortByDirectory(a: RepoProjectConfig, b: RepoProjectConfig): number {
@@ -282,7 +280,7 @@ export function findProjectsFromPath(
   path: string
 ): RepoProjectConfig[] {
   const normalizedPath = normalizePath(path);
-  return projects
+  const matches = projects
     .slice()
     .sort(sortByDirectory)
     .filter(project => {
@@ -295,14 +293,9 @@ export function findProjectsFromPath(
         normalizedPath.startsWith(`${project.directory}/`)
       );
     });
-}
-
-/**
- * TODO: remove
- */
-export function findProjectFromPath(
-  projects: RepoProjectConfig[],
-  path: string
-): RepoProjectConfig | undefined {
-  return findProjectsFromPath(projects, path)[0];
+  // If there are multiple matches, we only want the most relevant
+  // selections (with the deepest directory depth), so pick the first
+  // one and filter on those matches.
+  const firstMatch = matches[0];
+  return matches.filter(match => match.directory === firstMatch.directory);
 }